requirements.txt
# ----------------------
# Flask
# ----------------------
flask
flask-script
WTForms
mongoengine
flask_mongoengine
flask-sqlalchemy
psycopg2
flask-Login
flask-OpenID
Flask-Migrate
flask-mail
# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gunicorn
